fix:  extracted the wrpper object
