+++ test/unit/extensionManager/ExtensionManagerFake.php
@@ -1,0 +1,51 @@
+<?php
+
+/**
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU General Public License
+ * as published by the Free Software Foundation; under version 2
+ * of the License (non-upgradable).
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+ *
+ * Copyright (c) 2021 (original work) Open Assessment Technologies SA (under the project TAO-PRODUCT);
+ *
+ */
+
+declare(strict_types=1);
+
+namespace oat\tao\test\unit\extensionManager;
+
+use tao_actions_ExtensionsManager;
+
+class ExtensionManagerFake extends tao_actions_ExtensionsManager
+{
+    public function __construct()
+    {
+        // to avoid to set header because body is already sent
+    }
+
+    public function setView($path, $extensionID = null)
+    {
+        // avoid to call Template::getTemplate
+    }
+
+    public function hasRequestParameter($name)
+    {
+        // to intercept the flow and avoid to load common_request
+        throw new ExpectationFailedException('HTTP request cannot be handled by unit test');
+    }
+
+    public function getRequestParameter($name)
+    {
+        // to intercept the flow and avoid to load common_request
+        throw new ExpectationFailedException('HTTP request cannot be handled by unit test');
+    }
+}
